chore: refactor functions into files FUI-1754
